Combine meshgen modules
