+++ UM2N/generator/__init__.py
@@ -3,7 +3,6 @@ from .equation_solver import *  # noqa
 from .rand_source_generator import *  # noqa
 from .equation_generator import *  # noqa
 
-from .polymesh import *  # noqa
-from .squaremesh import *  # noqa
+from .meshgen import *  # noqa
 from .burgers_solver import *  # noqa
 from .swirl_solver import *  # noqa
